Remove `jaxrpc-api` Maven Distribution
`jaxrpc-api` is not maintained in this repo (in fact, this repo has an archived copy of `jaxrpc-api-1.1`), so it doesn't make sense for this repo to publish it at all. Instead, the POM indicates that https://github.com/javaee/javax.xml.rpc is the authoritative source of the archive. And THAT repo indicates https://github.com/eclipse-ee4j/jax-rpc-api is now the repo for it. Either way, it should not be published from here.